börsi ja fixeeritud paketid on vadeldavad ja arvutab hinna välja. Juurde lisan veel universaal paketid ja kodumasinad

+++ ElektriRakendus/KodumasinaKasutus.vb
@@ -5,7 +5,7 @@ Public Class KodumasinaKasutus
     Inherits Kalkulaator
     Dim Elktrikogus As Double
     Dim Kulu As Double
-    Public Sub New(ByVal PaketHind As Double, ByVal kasutus As Double, ByVal Aeg As Double)
+    Public Sub New(ByVal PaketHind As Decimal, ByVal kasutus As Double, ByVal Aeg As Double)
         ' Baasklassi konstruktori väljakutse
         MyBase.New("KodumasinaKasutus", PaketHind, kasutus, Aeg)
     End Sub
